kernelci.runtime: rename from kernelci.lab
Rename kernelci.runtime as kernelci.lab and update all areas of the
code including unit tests accordingly.
