+++ kernelci/runtime/lava/lava_xmlrpc.py
@@ -9,7 +9,8 @@
 
 import xmlrpc.client
 import urllib.parse
-from kernelci.lab.lava import LavaRuntime
+
+from kernelci.runtime.lava import LavaRuntime
 
 DEVICE_ONLINE_STATUS = ['idle', 'running', 'reserved']
 
@@ -17,7 +18,7 @@ DEVICE_ONLINE_STATUS = ['idle', 'running', 'reserved']
 class LAVA(LavaRuntime):
     """Interface to a LAVA lab
 
-    This implementation of kernelci.lab.LabAPI is to communicate with LAVA
+    This implementation of kernelci.runtime.LabAPI is to communicate with LAVA
     labs.  It can retrieve some information such as the list of devices and
     their online status, generate and submit jobs with callback parameters.
     One special thing it can deal with is job priorities, which is only
